+++ src/main/java/baseball/exception/DuplicateNumberException.java
@@ -1,4 +1,4 @@
-package baseball.console.exception;
+package baseball.exception;
 
 public class DuplicateNumberException extends IllegalNumberInputException {
