+++ src/main/java/baseball/exception/EachInputNumberOutOfRangeException.java
@@ -1,4 +1,4 @@
-package baseball.console.exception;
+package baseball.exception;
 
 public class EachInputNumberOutOfRangeException extends IllegalNumberInputException {
