+++ src/main/java/baseball/exception/IllegalNumberInputException.java
@@ -1,4 +1,4 @@
-package baseball.console.exception;
+package baseball.exception;
 
 public class IllegalNumberInputException extends RuntimeException {
