+++ src/main/java/baseball/exception/InputNumberFormatException.java
@@ -1,4 +1,4 @@
-package baseball.console.exception;
+package baseball.exception;
 
 public class InputNumberFormatException extends IllegalNumberInputException {
