+++ src/main/java/baseball/exception/InputNumberOutOfRangeException.java
@@ -1,4 +1,4 @@
-package baseball.console.exception;
+package baseball.exception;
 
 public class InputNumberOutOfRangeException extends IllegalNumberInputException {
