+++ src/main/java/baseball/exception/InvalidStatusCodeException.java
@@ -1,4 +1,4 @@
-package baseball.console.exception;
+package baseball.exception;
 
 public class InvalidStatusCodeException extends IllegalNumberInputException {
